* commit 'aee93922244fdaf3449e8f2e9bce799d60c2a134':
  sf: use flash->cmd_len instead of the sizeof(cmd)
  add flash deviced id to support the sixth byte
  sf: sf_params: Fix S25FS512S flash erase size

+++ drivers/mtd/spi/sf_internal.h
@@ -153,7 +153,7 @@ int sst_write_bp(struct spi_flash *flash, u32 offset, size_t len,
 struct spi_flash_params {
 	const char *name;
 	u32 jedec;
-	u16 ext_jedec;
+	u32 ext_jedec;
 	u32 sector_size;
 	u32 nr_sectors;
 	u8 e_rd_cmd;
